refactor: Adapter 파일 명 변경 및 onClick 이벤트 추가
